Renaming WegasScriptEditor file to TempScriptEditor
