+++ wegas-app/src/main/node/wegas-react/src/Editor/Components/FormView/PageLoaderSelect.tsx
@@ -17,7 +17,7 @@ import {
 import { createScript } from '../../../Helper/wegasEntites';
 import { MessageString } from '../MessageString';
 import { SrcEditorLanguages } from '../ScriptEditors/editorHelpers';
-import { TempScriptEditor } from '../ScriptEditors/WegasScriptEditor';
+import { TempScriptEditor } from '../ScriptEditors/TempScriptEditor';
 import { CommonView, CommonViewContainer } from './commonView';
 import { Labeled, LabeledView } from './labeled';
 import { scriptEditStyle } from './Script/Script';
