+++ wegas-app/src/main/node/wegas-react/src/Editor/Components/PlayLocal.tsx
@@ -11,7 +11,7 @@ import { themeVar } from '../../Components/Theme/ThemeVars';
 import { Toolbar } from '../../Components/Toolbar';
 import { defaultPadding, flex } from '../../css/classes';
 import { shallowIs } from '../../Helper/shallowIs';
-import { TempScriptEditor } from './ScriptEditors/WegasScriptEditor';
+import { TempScriptEditor } from './ScriptEditors/TempScriptEditor';
 
 const container = css({
   width: '100%',
