+++ wegas-app/src/main/node/wegas-react/src/Editor/Components/ScriptEditors/TempScriptEditor.tsx
